+++ event_reporter/reporter/application_errors_parser.go
@@ -39,29 +39,40 @@ func parseApplicationSyncResultErrorsFromConditions(status appv1.ApplicationStat
 		return errs
 	}
 	for _, cnd := range status.Conditions {
-		if !strings.Contains(strings.ToLower(cnd.Type), "error") {
-			continue
-		}
-
-		lastSeen := metav1.Now()
-		if cnd.LastTransitionTime != nil {
-			lastSeen = *cnd.LastTransitionTime
-		}
-
 		if (strings.Contains(cnd.Message, syncTaskUnsuccessfullErrorMessage) || strings.Contains(cnd.Message, syncTaskNotValidErrorMessage)) && status.OperationState != nil && status.OperationState.SyncResult != nil && status.OperationState.SyncResult.Resources != nil {
 			resourcesSyncErrors := parseAggregativeResourcesSyncErrors(status.OperationState.SyncResult.Resources)
 
 			errs = append(errs, resourcesSyncErrors...)
-		} else {
+			continue
+		}
+
+		if level := getConditionLevel(cnd); level != "" {
 			errs = append(errs, &events.ObjectError{
 				Type:     "sync",
-				Level:    "error",
+				Level:    level,
 				Message:  cnd.Message,
-				LastSeen: lastSeen,
+				LastSeen: getConditionTime(cnd),
 			})
 		}
 	}
 	return errs
+}
+
+func getConditionLevel(cnd appv1.ApplicationCondition) string {
+	if cnd.IsWarning() {
+		return "warning"
+	}
+	if cnd.IsError() {
+		return "error"
+	}
+	return ""
+}
+
+func getConditionTime(cnd appv1.ApplicationCondition) metav1.Time {
+	if cnd.LastTransitionTime != nil {
+		return *cnd.LastTransitionTime
+	}
+	return metav1.Now()
 }
 
 func parseResourceSyncResultErrors(rs *appv1.ResourceStatus, os *appv1.OperationState) []*events.ObjectError {
@@ -133,7 +144,7 @@ func parseAggregativeHealthErrors(rs *appv1.ResourceStatus, apptree *appv1.Appli
 			}
 
 			if addReference {
-				newErr.SourceReference = events.ErrorSourceReference{
+				newErr.SourceReference = &events.ErrorSourceReference{
 					Group:     rs.Group,
 					Version:   rs.Version,
 					Kind:      rs.Kind,